requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.2
django-ninja==1.3.0
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
